+++ frontend/src/components/StatsChart.js
